remove atomicfile
