+++ asdf/_tests/test_api.py
@@ -112,7 +112,7 @@ def test_atomic_write(tmp_path, small_tree):
     ff = asdf.AsdfFile(small_tree)
     ff.write_to(tmpfile)
 
-    with asdf.open(tmpfile, mode="r") as ff:
+    with asdf.open(tmpfile, mode="r", copy_arrays=True, lazy_load=False) as ff:
         ff.write_to(tmpfile)
